fix(tokens): fetch token lists on demand
