+++ apps/kyberswap-interface/src/components/LimitOrder/OrderBook/index.tsx
@@ -16,8 +16,8 @@ import { formatPriceInputValue } from 'components/LimitOrder/utils'
 import RefetchIndicator from 'components/RefetchIndicator'
 import RefreshLoading from 'components/RefreshLoading'
 import { useActiveWeb3React } from 'hooks'
-import { useStableCoins } from 'hooks/Tokens'
 import { useBaseTradeInfoLimitOrder } from 'hooks/useBaseTradeInfo'
+import { useStableCoins } from 'hooks/useTokens'
 import { getDefaultRevertPrice as getDefaultInvertPrice } from 'pages/Earns/utils'
 import { useLimitState } from 'state/limit/hooks'
 import { cn } from 'utils/cn'
