+++ apps/kyberswap-interface/src/components/Menu/hooks/useFaucetReward.ts
@@ -7,7 +7,7 @@ import { NotificationType } from 'components/Announcement/type'
 import { REWARD_SERVICE_API } from 'constants/env'
 import { NativeCurrencies } from 'constants/tokens'
 import { useActiveWeb3React } from 'hooks'
-import { useAllTokens } from 'hooks/Tokens'
+import { useAllTokens } from 'hooks/useTokens'
 import { useNotify } from 'state/application/hooks'
 import { getNativeTokenLogo, getTokenLogoURL, isAddress } from 'utils'
 import { filterTokens } from 'utils/filtering'
