+++ apps/kyberswap-interface/src/components/WalletPopup/Transactions/index.tsx
@@ -12,8 +12,8 @@ import Tab from 'components/WalletPopup/Transactions/Tab'
 import TransactionItem from 'components/WalletPopup/Transactions/TransactionItem'
 import { NUMBERS } from 'components/WalletPopup/Transactions/helper'
 import { useActiveWeb3React } from 'hooks'
-import { fetchListTokenByAddresses, findCacheToken, useIsLoadedTokenDefault } from 'hooks/Tokens'
 import { isSupportKyberDao } from 'hooks/kyberdao'
+import { fetchListTokenByAddresses, findCacheToken, useIsLoadedTokenDefault } from 'hooks/useTokens'
 import { useAppDispatch } from 'state/hooks'
 import { clearAllPendingTransactions } from 'state/transactions/actions'
 import { useSortRecentTransactions } from 'state/transactions/hooks'
