+++ apps/kyberswap-interface/src/pages/Earns/PositionDetail/InformationTab.tsx
@@ -20,7 +20,7 @@ import { APP_PATHS } from 'constants/index'
 import { NETWORKS_INFO } from 'constants/networks'
 import { NativeCurrencies } from 'constants/tokens'
 import { useActiveWeb3React } from 'hooks'
-import { useStableCoins } from 'hooks/Tokens'
+import { useStableCoins } from 'hooks/useTokens'
 import { timings } from 'pages/Earns/PoolExplorer/Filter'
 import LiquidityChart, { LiquidityChartSkeleton } from 'pages/Earns/PositionDetail/LiquidityChart'
 import PositionAprTooltip from 'pages/Earns/PositionDetail/PositionAprTooltip'
