+++ apps/kyberswap-interface/src/pages/Earns/hooks/useMerklRewards.tsx
@@ -4,8 +4,8 @@ import { useLazySearchTokensBySymbolQuery } from 'services/ksSetting'
 import { useGetMerklChainsQuery, useMerklRewardsQuery } from 'services/rewardMerkl'
 
 import { useActiveWeb3React } from 'hooks'
-import { fetchListTokenByAddresses } from 'hooks/Tokens'
 import useChainsConfig from 'hooks/useChainsConfig'
+import { fetchListTokenByAddresses } from 'hooks/useTokens'
 import useFilter from 'pages/Earns/UserPositions/useFilter'
 import { EarnChain } from 'pages/Earns/constants'
 import { ChainRewardInfo, ParsedPosition, TokenRewardInfo } from 'pages/Earns/types'
