+++ apps/kyberswap-interface/src/pages/ElasticLegacy/FarmLegacy.tsx
@@ -12,10 +12,10 @@ import { PROMM_FARM_ABI } from 'constants/abis'
 import { ELASTIC_BASE_FEE_UNIT, ZERO_ADDRESS } from 'constants/index'
 import { NativeCurrencies } from 'constants/tokens'
 import { useActiveWeb3React, useWeb3React } from 'hooks'
-import { useAllTokens } from 'hooks/Tokens'
 import { useProMMFarmSigningContract } from 'hooks/useContract'
 import { Position as SubgraphPosition, config, parsePosition } from 'hooks/useElasticLegacy'
 import useTheme from 'hooks/useTheme'
+import { useAllTokens } from 'hooks/useTokens'
 import { FeeTag } from 'pages/ElasticLegacy/PositionLegacy'
 import { useTokenPrices } from 'state/tokenPrices/hooks'
 import { useTransactionAdder } from 'state/transactions/hooks'
