+++ apps/kyberswap-interface/src/pages/ElasticSnapshot/components/InstantClaimModal.tsx
@@ -16,8 +16,8 @@ import Modal from 'components/Modal'
 import { wagmiConfig } from 'components/Web3Provider'
 import { NETWORKS_INFO } from 'constants/networks'
 import { useActiveWeb3React, useWeb3React } from 'hooks'
-import { useAllTokens } from 'hooks/Tokens'
 import { useReadingContract } from 'hooks/useContract'
+import { useAllTokens } from 'hooks/useTokens'
 import { useChangeNetwork } from 'hooks/web3/useChangeNetwork'
 import InstantAbi from 'pages/ElasticSnapshot/data/abis/instantClaimAbi.json'
 import avalanche from 'pages/ElasticSnapshot/data/instant/avalanche.json'
