+++ apps/kyberswap-interface/src/pages/ElasticSnapshot/index.tsx
@@ -11,9 +11,9 @@ import { PoolsPageWrapper } from 'components/PageWrappers'
 import Tabs from 'components/Tabs'
 import { MouseoverTooltip } from 'components/Tooltip'
 import { useActiveWeb3React } from 'hooks'
-import { useAllTokens } from 'hooks/Tokens'
 import { NETWORKS_INFO } from 'hooks/useChainsConfig'
 import useTheme from 'hooks/useTheme'
+import { useAllTokens } from 'hooks/useTokens'
 import TreasuryGrantAndInstantClaim from 'pages/ElasticSnapshot/components/TreasuryGrantAndInstantClaim'
 import Vesting, { VestingInterface } from 'pages/ElasticSnapshot/components/Vesting'
 import poolsByCategoriesRaw from 'pages/ElasticSnapshot/data/category.json'
