+++ apps/kyberswap-interface/src/pages/ProAmmPool/PositionListItem.tsx
@@ -14,9 +14,9 @@ import { RowBetween } from 'components/Row'
 import { MouseoverTooltip } from 'components/Tooltip'
 import { APP_PATHS, PROMM_ANALYTICS_URL } from 'constants/index'
 import { useActiveWeb3React } from 'hooks'
-import { useToken } from 'hooks/Tokens'
 import useIsTickAtLimit from 'hooks/useIsTickAtLimit'
 import { usePool } from 'hooks/usePools'
+import { useToken } from 'hooks/useTokens'
 import useTracking, { TRACKING_EVENT_TYPE } from 'hooks/useTracking'
 import ContentLoader from 'pages/ProAmmPool/ContentLoader'
 import { useTokenPrices } from 'state/tokenPrices/hooks'
