+++ apps/kyberswap-interface/src/pages/RemoveLiquidity/index.tsx
@@ -8,7 +8,7 @@ import LiquidityProviderMode from 'components/LiquidityProviderMode'
 import { AddRemoveTabs, LiquidityAction } from 'components/NavigationTabs'
 import { MinimalPositionCard } from 'components/PositionCard'
 import { useActiveWeb3React } from 'hooks'
-import { useCurrency } from 'hooks/Tokens'
+import { useCurrency } from 'hooks/useTokens'
 import useTracking, { TRACKING_EVENT_TYPE } from 'hooks/useTracking'
 import { useDerivedBurnInfo } from 'state/burn/hooks'
 import { useCurrencyConvertedToNative } from 'utils/dmm'
