+++ apps/kyberswap-interface/src/pages/SwapV3/index.tsx
@@ -17,9 +17,9 @@ import { Container, InfoComponentsWrapper, PageWrapper, SwapFormWrapper } from '
 import { APP_PATHS } from 'constants/index'
 import { PRICE_CHART_QUOTE_TOKEN_BY_CHAIN } from 'constants/tokens'
 import { useActiveWeb3React } from 'hooks'
-import { useAllTokens } from 'hooks/Tokens'
 import { NETWORKS_INFO } from 'hooks/useChainsConfig'
 import useParsedQueryString from 'hooks/useParsedQueryString'
+import { useAllTokens } from 'hooks/useTokens'
 import CrossChainSwap from 'pages/CrossChainSwap'
 import { CrossChainSwapSources } from 'pages/CrossChainSwap/components/CrossChainSwapSources'
 import QuoteSteps from 'pages/CrossChainSwap/components/QuoteSteps'
