+++ apps/kyberswap-interface/src/state/burn/proamm/hooks.tsx
@@ -4,8 +4,8 @@ import { Trans } from '@lingui/macro'
 import { ReactNode, useCallback, useMemo } from 'react'
 
 import { useActiveWeb3React } from 'hooks'
-import { useToken } from 'hooks/Tokens'
 import { usePool } from 'hooks/usePools'
+import { useToken } from 'hooks/useTokens'
 import { AppState } from 'state'
 import { useAppDispatch, useAppSelector } from 'state/hooks'
 import { tryParseAmount } from 'state/swap/hooks'
