+++ apps/kyberswap-interface/src/state/lists/actions.ts
@@ -1,7 +1,7 @@
 import { ChainId } from '@kyberswap/ks-sdk-core'
 import { ActionCreatorWithPayload, createAction } from '@reduxjs/toolkit'
 
-import { TokenMap } from 'hooks/Tokens'
+import { TokenMap } from 'hooks/useTokens'
 
 export const setTokenList: ActionCreatorWithPayload<{ chainId: ChainId; tokenList: TokenMap }> =
   createAction('lists/setTokenList')
