+++ apps/kyberswap-interface/src/state/swap/hooks.ts
@@ -7,8 +7,8 @@ import { useCheckPairQuery } from 'services/marketOverview'
 import { APP_PATHS, PAIR_CATEGORY } from 'constants/index'
 import { DEFAULT_OUTPUT_TOKEN_BY_CHAIN, NativeCurrencies } from 'constants/tokens'
 import { useActiveWeb3React } from 'hooks'
-import { useAllTokens, useCurrencyV2, useStableCoins } from 'hooks/Tokens'
 import { NETWORKS_INFO } from 'hooks/useChainsConfig'
+import { useAllTokens, useCurrencyV2, useStableCoins } from 'hooks/useTokens'
 import { useAppSelector } from 'state/hooks'
 import { AppDispatch, AppState } from 'state/index'
 import { Field, resetSelectCurrency, setRecipient, setTrade, typeInput } from 'state/swap/actions'
